backend/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
fastapi-cors